Move Swift compiler version parsing from `SkipUnless.swift` to `Toolchain`
This allows us to inspect the Swift version and fix up compiler arguments from SwiftPM to account for an older version of SwiftPM being invoked using `swift build` than the one that’s bundled in-process with SourceKit-LSP.
